conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
basemap=1.4.0=py310hbda83bc_0
basemap-data=1.4.0=py310hca03da5_0
blas=1.0=openblas
bottleneck=1.3.7=py310hbda83bc_0
brotli=1.0.9=h80987f9_8
brotli-bin=1.0.9=h80987f9_8
bzip2=1.0.8=h80987f9_6
c-ares=1.19.1=h80987f9_0
ca-certificates=2024.7.2=hca03da5_0
certifi=2024.8.30=py310hca03da5_0
charset-normalizer=3.3.2=pypi_0
contourpy=1.2.0=py310h48ca7d4_0
cycler=0.11.0=pyhd3eb1b0_0
et_xmlfile=1.1.0=py310hca03da5_0
fonttools=4.51.0=py310h80987f9_0
freetype=2.12.1=h1192e45_0
geos=3.9.1=hc377ac9_1
idna=3.8=pypi_0
jpeg=9e=h80987f9_3
kiwisolver=1.4.4=py310h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h80987f9_8
libbrotlidec=1.0.9=h80987f9_8
libbrotlienc=1.0.9=h80987f9_8
libcurl=8.9.1=h3e2b118_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libnghttp2=1.57.0=h62f6fdd_0
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libssh2=1.11.0=h3e2b118_0
libtiff=4.5.1=h313beb8_0
libwebp-base=1.3.2=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
matplotlib-base=3.8.4=py310h46d7db6_0
ncurses=6.4=h313beb8_0
numexpr=2.8.7=py310hecc3335_0
numpy=2.1.1=pypi_0
numpy-base=1.26.4=py310ha9811e2_0
openjpeg=2.5.2=h54b8e55_0
openpyxl=3.1.5=py310h80987f9_0
openssl=3.0.15=h80987f9_0
packaging=24.1=pypi_0
panda=0.3.1=pypi_0
pandas=2.2.2=py310h313beb8_0
pillow=10.4.0=py310h80987f9_0
pip=24.2=py310hca03da5_0
plotly=5.24.0=pypi_0
proj=9.3.1=h805f6d4_0
pybind11-abi=4=hd3eb1b0_1
pyparsing=3.0.9=py310hca03da5_0
pyproj=3.6.1=py310h6070227_0
pyshp=2.3.1=py310hca03da5_0
python=3.10.14=hb885b13_1
python-dateutil=2.9.0post0=py310hca03da5_2
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py310hca03da5_0
readline=8.2=h1a28f6b_0
requests=2.32.3=pypi_0
scipy=1.13.1=py310hd336fd7_0
setuptools=72.1.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h80987f9_0
tenacity=9.0.0=pypi_0
tk=8.6.14=h6ba3021_0
tzdata=2024.1=pypi_0
unicodedata2=15.1.0=py310h80987f9_0
urllib3=2.2.2=pypi_0
wheel=0.43.0=py310hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
zstd=1.5.5=hd90d995_2
